refactor(fan): extract schedule and profile logic into separate modules
Split FanControlService (1305 lines) into focused modules:
- fan_schedule.py: FanScheduleService with schedule CRUD and time-based curve resolution
- fan_profiles.py: FanProfileService with profile CRUD and preset application
- fan_control.py: core monitoring, PWM control, and hysteresis (908 lines)

Backward-compatible delegating methods kept on FanControlService.
Also fixes unbound 'sensors' variable and Optional[Settings] type hint.
