+++ backend/app/services/power/fan_control.py
@@ -14,12 +14,13 @@ from dataclasses import dataclass
 from datetime import datetime, timezone
 from typing import Dict, List, Optional, Tuple
 
-from sqlalchemy import select, desc
-from sqlalchemy.orm import Session
+from sqlalchemy import select, desc, func
 
 from app.core.config import Settings
-from app.models.fans import FanConfig, FanSample, FanScheduleEntry, FanCurveProfile
+from app.models.fans import FanConfig, FanSample
 from app.schemas.fans import FanMode, FanCurvePoint
+from app.services.power.fan_schedule import FanScheduleService
+from app.services.power.fan_profiles import FanProfileService
 
 logger = logging.getLogger(__name__)
 
@@ -133,11 +134,23 @@ class FanControlService:
         self._is_running = False
         self._use_linux_backend = False
         self._hysteresis_state: Dict[str, HysteresisState] = {}  # Track hysteresis per fan
+        self._schedule = FanScheduleService(db_session_factory)
+        self._profiles = FanProfileService(db_session_factory)
 
         FanControlService._instance = self
 
+    @property
+    def schedule(self) -> FanScheduleService:
+        """Access the fan schedule service."""
+        return self._schedule
+
+    @property
+    def profiles(self) -> FanProfileService:
+        """Access the fan profile service."""
+        return self._profiles
+
     @classmethod
-    async def get_instance(cls, config: Settings = None, db_session_factory=None) -> "FanControlService":
+    async def get_instance(cls, config: Optional[Settings] = None, db_session_factory=None) -> "FanControlService":
         """Get singleton instance."""
         async with cls._lock:
             if cls._instance is None:
@@ -219,6 +232,7 @@ class FanControlService:
 
         # Determine best CPU sensor for auto-correction
         cpu_sensor_id: Optional[str] = None
+        sensors: List[TempSensorData] = []
         try:
             sensors = await self._backend.get_available_temp_sensors()
             for s in sensors:
@@ -231,12 +245,9 @@ class FanControlService:
         # Build set of known CPU sensor IDs for checking existing configs
         cpu_sensor_ids: set = set()
         if cpu_sensor_id:
-            try:
-                for s in sensors:
-                    if s.is_cpu_sensor:
-                        cpu_sensor_ids.add(s.sensor_id)
-            except Exception:
-                pass
+            for s in sensors:
+                if s.is_cpu_sensor:
+                    cpu_sensor_ids.add(s.sensor_id)
 
         with self.db_session_factory() as db:
             for fan in fans:
@@ -348,7 +359,7 @@ class FanControlService:
                         else:
                             # Resolve curve: scheduled mode uses time-based curve, auto uses default
                             if FanMode(config.mode) == FanMode.SCHEDULED:
-                                curve_points, _ = self._resolve_active_curve(
+                                curve_points, _ = self._schedule.resolve_active_curve(
                                     fan.fan_id, config.curve_json, db
                                 )
                             else:
@@ -469,80 +480,6 @@ class FanControlService:
         # PWM unchanged
         return state.last_pwm
 
-    @staticmethod
-    def _time_in_window(current_minutes: int, start_minutes: int, end_minutes: int) -> bool:
-        """
-        Check if current time (in minutes since midnight) falls within a window.
-
-        Supports overnight windows (e.g. 22:00-06:00).
-
-        Args:
-            current_minutes: Current time as minutes since midnight (0-1439)
-            start_minutes: Window start as minutes since midnight
-            end_minutes: Window end as minutes since midnight
-
-        Returns:
-            True if current time is within the window
-        """
-        if start_minutes <= end_minutes:
-            # Normal window (e.g. 08:00-18:00)
-            return start_minutes <= current_minutes < end_minutes
-        else:
-            # Overnight window (e.g. 22:00-06:00)
-            return current_minutes >= start_minutes or current_minutes < end_minutes
-
-    @staticmethod
-    def _parse_time_to_minutes(time_str: str) -> int:
-        """Parse HH:MM string to minutes since midnight."""
-        parts = time_str.split(':')
-        return int(parts[0]) * 60 + int(parts[1])
-
-    def _resolve_active_curve(
-        self, fan_id: str, default_curve_json: Optional[str], db: Session
-    ) -> Tuple[List[dict], Optional[FanScheduleEntry]]:
-        """
-        Find the active schedule entry for the current time.
-
-        Args:
-            fan_id: Fan identifier
-            default_curve_json: Default curve from FanConfig
-            db: Database session
-
-        Returns:
-            Tuple of (curve_points list, active FanScheduleEntry or None)
-        """
-        now = datetime.now()
-        current_minutes = now.hour * 60 + now.minute
-
-        entries = db.execute(
-            select(FanScheduleEntry)
-            .where(FanScheduleEntry.fan_id == fan_id)
-            .where(FanScheduleEntry.is_enabled == True)
-            .order_by(FanScheduleEntry.priority.asc())
-        ).scalars().all()
-
-        for entry in entries:
-            start = self._parse_time_to_minutes(entry.start_time)
-            end = self._parse_time_to_minutes(entry.end_time)
-            if self._time_in_window(current_minutes, start, end):
-                # If entry references a profile, use the profile's curve
-                if entry.profile_id is not None:
-                    profile = db.execute(
-                        select(FanCurveProfile).where(FanCurveProfile.id == entry.profile_id)
-                    ).scalar_one_or_none()
-                    if profile:
-                        curve = json.loads(profile.curve_json) if profile.curve_json else []
-                        if len(curve) >= 2:
-                            return curve, entry
-                # Fallback to entry's inline curve
-                curve = json.loads(entry.curve_json) if entry.curve_json else []
-                if len(curve) >= 2:
-                    return curve, entry
-
-        # Fallback to default curve
-        default_curve = json.loads(default_curve_json) if default_curve_json else []
-        return default_curve, None
-
     async def _persist_samples(self):
         """Persist buffered samples to database."""
         if not self._sample_buffer:
@@ -613,7 +550,7 @@ class FanControlService:
 
                     # Add schedule info for scheduled mode
                     if config.mode == FanMode.SCHEDULED.value:
-                        _, active_entry = self._resolve_active_curve(
+                        _, active_entry = self._schedule.resolve_active_curve(
                             fan.fan_id, config.curve_json, db
                         )
                         if active_entry:
@@ -786,381 +723,49 @@ class FanControlService:
             logger.info("Switched to dev fan control backend")
             return True, False
 
-    async def apply_preset(self, fan_id: str, preset_name: str) -> Tuple[bool, List[FanCurvePoint]]:
-        """
-        Apply a preset curve to a fan.
+    # --- Delegating methods for backward compatibility ---
 
-        Looks up system profiles from DB first, falls back to CURVE_PRESETS dict.
+    async def get_schedule_entries(self, fan_id: str):
+        return await self._schedule.get_schedule_entries(fan_id)
 
-        Args:
-            fan_id: Fan identifier
-            preset_name: Preset name (silent, balanced, performance)
+    async def create_schedule_entry(self, fan_id: str, name: str, start_time: str, end_time: str, **kwargs):
+        return await self._schedule.create_schedule_entry(fan_id, name, start_time, end_time, **kwargs)
 
-        Returns:
-            (success, curve_points)
-        """
-        # Try DB system profile first
-        curve_points: Optional[List[FanCurvePoint]] = None
-        with self.db_session_factory() as db:
-            profile = db.execute(
-                select(FanCurveProfile)
-                .where(FanCurveProfile.name == preset_name)
-                .where(FanCurveProfile.is_system == True)
-            ).scalar_one_or_none()
-            if profile:
-                points = json.loads(profile.curve_json)
-                curve_points = [FanCurvePoint(temp=p["temp"], pwm=p["pwm"]) for p in points]
-
-        # Fallback to hardcoded presets
-        if curve_points is None:
-            from app.schemas.fans import CURVE_PRESETS
-            if preset_name not in CURVE_PRESETS:
-                logger.warning(f"Unknown preset: {preset_name}")
-                return False, []
-            preset_points = CURVE_PRESETS[preset_name]
-            curve_points = [FanCurvePoint(temp=p["temp"], pwm=p["pwm"]) for p in preset_points]
-
-        success = await self.update_fan_curve(fan_id, curve_points)
-        if success:
-            # Clear hysteresis state when curve changes
-            if fan_id in self._hysteresis_state:
-                del self._hysteresis_state[fan_id]
-            logger.info(f"Applied {preset_name} preset to {fan_id}")
-
-        return success, curve_points
-
-    async def get_schedule_entries(self, fan_id: str) -> List[FanScheduleEntry]:
-        """Get all schedule entries for a fan."""
-        with self.db_session_factory() as db:
-            entries = db.execute(
-                select(FanScheduleEntry)
-                .where(FanScheduleEntry.fan_id == fan_id)
-                .order_by(FanScheduleEntry.priority.asc(), FanScheduleEntry.start_time.asc())
-            ).scalars().all()
-            # Detach from session before returning
-            for entry in entries:
-                db.expunge(entry)
-            return list(entries)
-
-    async def create_schedule_entry(
-        self, fan_id: str, name: str, start_time: str, end_time: str,
-        curve_points: Optional[List[FanCurvePoint]] = None,
-        priority: int = 0, is_enabled: bool = True,
-        profile_id: Optional[int] = None,
-    ) -> Optional[FanScheduleEntry]:
-        """
-        Create a new schedule entry for a fan.
-
-        Returns None if max entries (8) reached.
-        """
-        with self.db_session_factory() as db:
-            # Check max entries
-            count = db.execute(
-                select(func.count()).select_from(
-                    select(FanScheduleEntry)
-                    .where(FanScheduleEntry.fan_id == fan_id)
-                    .subquery()
-                )
-            ).scalar() or 0
-
-            if count >= 8:
-                return None
-
-            curve_json = None
-            if curve_points:
-                curve_json = json.dumps([p.model_dump() for p in curve_points])
-
-            entry = FanScheduleEntry(
-                fan_id=fan_id,
-                name=name,
-                start_time=start_time,
-                end_time=end_time,
-                curve_json=curve_json,
-                priority=priority,
-                is_enabled=is_enabled,
-                profile_id=profile_id,
-            )
-            db.add(entry)
-            db.commit()
-            db.refresh(entry)
-            db.expunge(entry)
-
-            logger.info(f"Created schedule entry '{name}' for {fan_id} ({start_time}-{end_time})")
-            return entry
-
-    async def update_schedule_entry(
-        self, fan_id: str, entry_id: int, **kwargs
-    ) -> Optional[FanScheduleEntry]:
-        """Update an existing schedule entry."""
-        with self.db_session_factory() as db:
-            entry = db.execute(
-                select(FanScheduleEntry)
-                .where(FanScheduleEntry.id == entry_id)
-                .where(FanScheduleEntry.fan_id == fan_id)
-            ).scalar_one_or_none()
-
-            if not entry:
-                return None
-
-            if 'name' in kwargs and kwargs['name'] is not None:
-                entry.name = kwargs['name']
-            if 'start_time' in kwargs and kwargs['start_time'] is not None:
-                entry.start_time = kwargs['start_time']
-            if 'end_time' in kwargs and kwargs['end_time'] is not None:
-                entry.end_time = kwargs['end_time']
-            if 'curve_points' in kwargs and kwargs['curve_points'] is not None:
-                entry.curve_json = json.dumps([p.model_dump() for p in kwargs['curve_points']])
-                entry.profile_id = None  # Clear profile when setting inline curve
-            if 'profile_id' in kwargs:
-                pid = kwargs['profile_id']
-                if pid is not None:
-                    entry.profile_id = pid
-                    entry.curve_json = None  # Clear inline curve when setting profile
-                else:
-                    entry.profile_id = None
-            if 'priority' in kwargs and kwargs['priority'] is not None:
-                entry.priority = kwargs['priority']
-            if 'is_enabled' in kwargs and kwargs['is_enabled'] is not None:
-                entry.is_enabled = kwargs['is_enabled']
-
-            db.commit()
-            db.refresh(entry)
-            db.expunge(entry)
-
-            logger.info(f"Updated schedule entry {entry_id} for {fan_id}")
-            return entry
+    async def update_schedule_entry(self, fan_id: str, entry_id: int, **kwargs):
+        return await self._schedule.update_schedule_entry(fan_id, entry_id, **kwargs)
 
     async def delete_schedule_entry(self, fan_id: str, entry_id: int) -> bool:
-        """Delete a schedule entry."""
-        with self.db_session_factory() as db:
-            entry = db.execute(
-                select(FanScheduleEntry)
-                .where(FanScheduleEntry.id == entry_id)
-                .where(FanScheduleEntry.fan_id == fan_id)
-            ).scalar_one_or_none()
+        return await self._schedule.delete_schedule_entry(fan_id, entry_id)
 
-            if not entry:
-                return False
+    async def get_active_schedule_entry(self, fan_id: str):
+        return await self._schedule.get_active_schedule_entry(fan_id)
 
-            db.delete(entry)
-            db.commit()
+    async def list_profiles(self):
+        return await self._profiles.list_profiles()
 
-            logger.info(f"Deleted schedule entry {entry_id} for {fan_id}")
-            return True
+    async def get_profile(self, profile_id: int):
+        return await self._profiles.get_profile(profile_id)
 
-    async def get_active_schedule_entry(self, fan_id: str) -> Tuple[Optional[FanScheduleEntry], Optional[FanScheduleEntry]]:
-        """
-        Get the currently active and next schedule entry for a fan.
+    async def create_profile(self, name: str, curve_points, description=None):
+        return await self._profiles.create_profile(name, curve_points, description)
 
-        Returns:
-            Tuple of (active_entry, next_entry)
-        """
-        with self.db_session_factory() as db:
-            config = db.execute(
-                select(FanConfig).where(FanConfig.fan_id == fan_id)
-            ).scalar_one_or_none()
-
-            if not config:
-                return None, None
-
-            _, active_entry = self._resolve_active_curve(fan_id, config.curve_json, db)
-
-            # Find next entry
-            now = datetime.now()
-            current_minutes = now.hour * 60 + now.minute
-            entries = db.execute(
-                select(FanScheduleEntry)
-                .where(FanScheduleEntry.fan_id == fan_id)
-                .where(FanScheduleEntry.is_enabled == True)
-                .order_by(FanScheduleEntry.start_time.asc())
-            ).scalars().all()
-
-            next_entry = None
-            for entry in entries:
-                start = self._parse_time_to_minutes(entry.start_time)
-                if start > current_minutes and entry != active_entry:
-                    next_entry = entry
-                    break
-
-            # Wrap around: if no future entry, next is the first one tomorrow
-            if next_entry is None and entries:
-                for entry in entries:
-                    if entry != active_entry:
-                        next_entry = entry
-                        break
-
-            # Detach from session
-            if active_entry:
-                db.expunge(active_entry)
-            if next_entry and next_entry is not active_entry:
-                db.expunge(next_entry)
-
-            return active_entry, next_entry
-
-    # --- Profile CRUD Methods ---
-
-    async def list_profiles(self) -> List[FanCurveProfile]:
-        """Get all fan curve profiles."""
-        with self.db_session_factory() as db:
-            profiles = db.execute(
-                select(FanCurveProfile).order_by(
-                    FanCurveProfile.is_system.desc(),
-                    FanCurveProfile.name.asc(),
-                )
-            ).scalars().all()
-            for p in profiles:
-                db.expunge(p)
-            return list(profiles)
-
-    async def get_profile(self, profile_id: int) -> Optional[FanCurveProfile]:
-        """Get a single profile by ID."""
-        with self.db_session_factory() as db:
-            profile = db.execute(
-                select(FanCurveProfile).where(FanCurveProfile.id == profile_id)
-            ).scalar_one_or_none()
-            if profile:
-                db.expunge(profile)
-            return profile
-
-    async def create_profile(
-        self, name: str, curve_points: List[FanCurvePoint], description: Optional[str] = None
-    ) -> Optional[FanCurveProfile]:
-        """
-        Create a new user curve profile.
-
-        Returns None if max user profiles (20) reached or name already exists.
-        """
-        with self.db_session_factory() as db:
-            # Check max user profiles
-            user_count = db.execute(
-                select(func.count()).select_from(
-                    select(FanCurveProfile)
-                    .where(FanCurveProfile.is_system == False)
-                    .subquery()
-                )
-            ).scalar() or 0
-            if user_count >= 20:
-                return None
-
-            # Check name uniqueness
-            existing = db.execute(
-                select(FanCurveProfile).where(FanCurveProfile.name == name)
-            ).scalar_one_or_none()
-            if existing:
-                return None
-
-            curve_json = json.dumps([p.model_dump() for p in curve_points])
-            profile = FanCurveProfile(
-                name=name,
-                description=description,
-                curve_json=curve_json,
-                is_system=False,
-            )
-            db.add(profile)
-            db.commit()
-            db.refresh(profile)
-            db.expunge(profile)
-
-            logger.info(f"Created curve profile '{name}'")
-            return profile
-
-    async def update_profile(
-        self, profile_id: int, **kwargs
-    ) -> Optional[FanCurveProfile]:
-        """
-        Update a curve profile.
-
-        Rejects name/is_system changes on system profiles.
-        Returns None if profile not found.
-        """
-        with self.db_session_factory() as db:
-            profile = db.execute(
-                select(FanCurveProfile).where(FanCurveProfile.id == profile_id)
-            ).scalar_one_or_none()
-
-            if not profile:
-                return None
-
-            if profile.is_system:
-                # System profiles: only allow curve_points and description updates
-                kwargs.pop('name', None)
-
-            if 'name' in kwargs and kwargs['name'] is not None:
-                # Check uniqueness
-                existing = db.execute(
-                    select(FanCurveProfile)
-                    .where(FanCurveProfile.name == kwargs['name'])
-                    .where(FanCurveProfile.id != profile_id)
-                ).scalar_one_or_none()
-                if existing:
-                    return None
-                profile.name = kwargs['name']
-
-            if 'description' in kwargs:
-                profile.description = kwargs['description']
-
-            if 'curve_points' in kwargs and kwargs['curve_points'] is not None:
-                profile.curve_json = json.dumps([p.model_dump() for p in kwargs['curve_points']])
-
-            db.commit()
-            db.refresh(profile)
-            db.expunge(profile)
-
-            logger.info(f"Updated curve profile {profile_id}")
-            return profile
+    async def update_profile(self, profile_id: int, **kwargs):
+        return await self._profiles.update_profile(profile_id, **kwargs)
 
     async def delete_profile(self, profile_id: int) -> bool:
-        """
-        Delete a curve profile.
+        return await self._profiles.delete_profile(profile_id)
 
-        Rejects deletion of system profiles. FK SET NULL handles schedule entries.
-        """
-        with self.db_session_factory() as db:
-            profile = db.execute(
-                select(FanCurveProfile).where(FanCurveProfile.id == profile_id)
-            ).scalar_one_or_none()
+    async def apply_profile_to_fan(self, fan_id: str, profile_id: int):
+        result = await self._profiles.apply_profile_to_fan(fan_id, profile_id)
+        if result[0] and fan_id in self._hysteresis_state:
+            del self._hysteresis_state[fan_id]
+        return result
 
-            if not profile:
-                return False
-
-            if profile.is_system:
-                return False
-
-            db.delete(profile)
-            db.commit()
-
-            logger.info(f"Deleted curve profile '{profile.name}' (id={profile_id})")
-            return True
-
-    async def apply_profile_to_fan(
-        self, fan_id: str, profile_id: int
-    ) -> Tuple[bool, List[FanCurvePoint]]:
-        """
-        Apply a profile's curve to a fan's FanConfig.
-
-        Copies the profile's curve points to the fan's curve_json.
-
-        Args:
-            fan_id: Fan identifier
-            profile_id: Profile ID
-
-        Returns:
-            (success, curve_points)
-        """
-        profile = await self.get_profile(profile_id)
-        if not profile:
-            return False, []
-
-        points = json.loads(profile.curve_json)
-        curve_points = [FanCurvePoint(temp=p["temp"], pwm=p["pwm"]) for p in points]
-
-        success = await self.update_fan_curve(fan_id, curve_points)
-        if success:
-            if fan_id in self._hysteresis_state:
-                del self._hysteresis_state[fan_id]
-            logger.info(f"Applied profile '{profile.name}' to {fan_id}")
-
-        return success, curve_points
+    async def apply_preset(self, fan_id: str, preset_name: str):
+        result = await self._profiles.apply_preset(fan_id, preset_name)
+        if result[0] and fan_id in self._hysteresis_state:
+            del self._hysteresis_state[fan_id]
+        return result
 
     async def update_fan_config(
         self,
@@ -1224,10 +829,6 @@ class FanControlService:
                 "emergency_temp_celsius": config.emergency_temp_celsius,
                 "temp_sensor_id": config.temp_sensor_id,
             }
-
-
-# Import for count function
-from sqlalchemy import func
 
 
 # Global service instance
